+++ project/LibDependencies.scala
@@ -5,8 +5,8 @@ import play.sbt.PlayImport.ws
 import sbt.ModuleID
 
 object LibDependencies {
-  val govukFrontendVersion: String = "4.5.0"
-  val hmrcFrontendVersion: String  = "5.28.0"
+  val govukFrontendVersion: String = "4.6.0"
+  val hmrcFrontendVersion: String  = "5.30.0"
   val playLanguageVersion: String  = "6.1.0"
 
   val compile: Seq[ModuleID] = dependencies(